backend/requirements.txt
# Ytili Backend Requirements
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.41.3
websockets==14.1

# Pydantic for data validation
pydantic==2.11.7
pydantic-settings==2.10.1

# Database
sqlalchemy==2.0.41
asyncpg==0.30.0

# Authentication & Security
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
cryptography==45.0.5

# HTTP client
httpx==0.28.1
requests==2.32.3

# Supabase integration
supabase==2.11.0

# Blockchain integration (Web3)
web3==7.12.1
eth-account==0.13.7

# Environment and configuration
python-dotenv==1.0.1
python-multipart==0.0.20

# Logging and monitoring
structlog==25.4.0

# Email validation
email-validator==2.2.0

# File handling
aiofiles==24.1.0
Pillow==11.3.0

# Payment processing (Stripe)
stripe==12.3.0

# Redis for caching
redis==6.2.0

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0

# AI Agent Dependencies
openai>=1.0.0  # For OpenRouter compatibility
tiktoken>=0.5.0  # Token counting for AI models
langchain>=0.1.0  # AI framework for conversation management
langchain-openai>=0.1.0  # OpenAI integration for LangChain

# OCR and Document Processing
pytesseract>=0.3.10  # OCR for medical document verification
opencv-python>=4.8.0  # Image processing for document analysis
Pillow>=10.0.0  # Image manipulation (already included but ensuring version)

# Natural Language Processing
spacy>=3.7.0  # NLP for Vietnamese medical terminology
transformers>=4.30.0  # Hugging Face transformers for local AI models

# Data Science and Analytics
pandas>=2.0.0  # Data analysis for predictive analytics
numpy>=1.24.0  # Numerical computing
scikit-learn>=1.3.0  # Machine learning for fraud detection

# Text Processing and Similarity
sentence-transformers>=2.2.0  # Semantic similarity for campaign matching
fuzzywuzzy>=0.18.0  # Fuzzy string matching for medical terms
python-Levenshtein>=0.21.0  # String distance calculations

# Async HTTP for AI API calls
aiohttp>=3.8.0  # Async HTTP client for OpenRouter API
